:truck: rename(file) : file 도메인 생성하면서 디렉토리 변경
